refactor: migracao da estrutura MVC para Feature-first (MVVM) (#28)
- Criada as pastas app/, shared/ e features/
- Arquivos movidos e divididos nas features certificados e gamificacao
- Adicionado lib/app/.gitkeep
- Atualizacao de todos os imports do projeto para rotas absolutas seguras
