+++ lib/shared/widgets/app_text.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import '../theme/app_theme.dart';
+import 'package:ifdex/shared/theme/app_theme.dart';
 
 class AppText extends StatelessWidget {
   final String text;
